% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createxyLabels.gen.digit.R
\name{createxyLabels.gen.digit}
\alias{createxyLabels.gen.digit}
\title{Create nice labels for Factor Maps and other maps (must be ggplot2 graphs), for both the x and y axes at the same time.}
\usage{
createxyLabels.gen.digit(
  x_axis = 1,
  y_axis = 2,
  lambda,
  tau,
  digit4lambda = 3,
  digit4tau = 0,
  axisName = "Dimension "
)
}
\arguments{
\item{x_axis}{Dimension number for x axis (default to 1) ;}

\item{y_axis}{Dimension number for y axis (default to 2) ;}

\item{lambda}{Eigenvalues ;}

\item{tau}{Percentage of explained variances ;}

\item{digit4lambda}{How many digits should be displayed for eigenvalues ;}

\item{digit4tau}{How many digits should be displayed for percentages of explained variances ;}

\item{axisName}{Name for the axes (e.g. "Dimension", "Component").}
}
\value{
x and y axis labels
}
\description{
Create nice labels for Factor Maps and other maps (must be ggplot2 graphs), for both the x and y axes at the same time.
}
\examples{
createxyLabels.gen.digit(lambda = 2:1, tau = 10*2:1)
}
