% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseCorrespondenceAnalysis.R
\docType{data}
\name{death2019}
\alias{death2019}
\title{The causes of death are categorized according to the
``International Statistical Classification of Diseases
and Related Health Problems''.
In this data set, columns are the causes of death
(corresponding to chapters in the ICD nomenclature);
rows are age groups going from the youngest
age group being from 1 to 5 years old to the oldest ages group being
100 and over.}
\description{
The causes of death are categorized according to the
``International Statistical Classification of Diseases
and Related Health Problems''.
In this data set, columns are the causes of death
(corresponding to chapters in the ICD nomenclature);
rows are age groups going from the youngest
age group being from 1 to 5 years old to the oldest ages group being
100 and over.
}
\references{
\url{https://www.cdc.gov/}
}
\keyword{data}
