% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pPMD.R
\name{pPMD}
\alias{pPMD}
\title{Projected PMD of a matrix.}
\usage{
pPMD(
  Data,
  k = 2L,
  rdsLeft = rep(1, k),
  rdsRight = rep(1, k),
  tol.si = .Machine$double.eps
)
}
\arguments{
\item{Data}{the rectangular matrix to decompose ;}

\item{k}{the desired rank of the singular decomposition (default to 2) ;}

\item{rdsLeft}{a vector of radiuses
(>0) of the $L_1$ or $L_G$ balls for each of the k left vectors ;}

\item{rdsRight}{a vector of radiuses
(>0) of the $L_1$ or $L_G$ balls for each of the k right vectors ;}

\item{tol.si}{tolerance for the sparsity}
}
\value{
Pseudo-singular vectors and values
}
\description{
Projected PMD of a matrix.
}
\examples{
X <- matrix(rnorm(20), 5, 4)
pPMD(X)
pPMD(
  X,
  k = 3L,
  rdsLeft = rep(0.5 * sqrt(5), 3),
  rdsRight = rep(0.5 * sqrt(4), 3))
}
\author{
Vincent Guillemot, Ju-Chi Yu
}
