% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPMD.R
\name{GPMD}
\alias{GPMD}
\title{Sparse Generalized Singular Value Decomposition}
\usage{
GPMD(
  X,
  LW,
  RW,
  k = 0,
  rdsLeft = rep(1, k),
  rdsRight = rep(1, k),
  tol.si = .Machine$double.eps
)
}
\arguments{
\item{X}{a (data) matrix;}

\item{LW}{PARAM_DESCRIPTION}

\item{RW}{PARAM_DESCRIPTION}

\item{k}{the desired rank of the singular decomposition, Default: 0}

\item{rdsLeft}{The radius (>0) of the
$L_1$ ball for each left vector, Default: rep(1, k)}

\item{rdsRight}{The radius (>0) of the $L_1$ balls for each right vector, Default: rep(1, k)}

\item{tol.si}{Tolerance for the computation of the Sparse Index, set by default to .Machine$double.eps}
}
\value{
Pseudo-singular vectors and values
}
\description{
Constrained SVD of a matrix (wrapper of PMA:::PMD).
}
\details{
DETAILS
}
\examples{
X <- matrix(rnorm(20), 5, 4)
GPMD(X, LW = diag(5), RW = diag(4))
}
\author{
Vincent Guillemot, Ju-Chi Yu
}
