% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseCorrespondenceAnalysis.R
\docType{package}
\name{sparseCorrespondenceAnalysis-package}
\alias{sparseCorrespondenceAnalysis}
\alias{sparseCorrespondenceAnalysis-package}
\title{Reproducible code for the paper "From Plain to Sparse Correspondence Analysis:
A Principal Component Analysis Approach". (Abdi et al., 2022)}
\description{
This package contains the reproducible code for the paper "From Plain to Sparse Correspondence Analysis: A Principal Component Analysis Approach". (Abdi et al., 2022)
}
\author{
\strong{Maintainer}: Vincent Guillemot \email{vincent.guillemot@pasteur.fr} (\href{https://orcid.org/0000-0002-7421-0655}{ORCID})

}
\keyword{internal}
